backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "DietMate – Your Smart Diet Companion. Plan meals, track calories, and balance nutrients effortlessly. Achieve your health goals with AI-powered recommendations. Eat smart, live better! 🍏"
authors = ["Przemyslaw Janiszewski", "Agnieszka Gilarska", "Krzysztof Trzcianowski"]
license = "MIT"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
pymongo = "*"
flask = "*"
flask-cors = "*"
python-dotenv = "*"
gunicorn = "*"
requests = "*"
groq = "*"
immutables = "*"
deep-translator = "*"
langdetect = "*"
PyJWT = "*"
BeautifulSoup4 = "*"
redis = "*"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"